version 1.0.0

+++ routes/api/tasks.js
@@ -1,13 +1,15 @@
-const express = require('express')
-const Router = express.Router()
+import express from 'express'
+import auth from '../../middleware/auth.js'
 
 // Task Modal
-const Task = require('../../models/Task')
+import Task from '../../models/Task.js'
+
+const Router = express.Router()
 
 // @route   GET api/tasks
 // @desc    Get all tasks
-// @access  Public
-Router.get('/', async (req, res) => {
+// @access  Private
+Router.get('/', auth, async (req, res) => {
     try {
         const tasks = await Task.find()
         tasks.sort((a, b) => {
@@ -15,37 +17,61 @@ Router.get('/', async (req, res) => {
         })
         res.json(tasks)
     } catch(err) {
-        res.status(400).json({ msg: err.toString() })
+        res.status(400).json({ msg: err.message })
     }
 })
 
 // @route   POST api/tasks
 // @desc    Create a task
-// @access  Public
-Router.post('/', async (req, res) => {
+// @access  Private
+Router.post('/', auth, async (req, res) => {
+    const {
+        title,
+        category,
+        description,
+        important
+    } = req.body
+
+    if(!title || !category || !description ||
+        typeof important === 'undefined') {
+        return res.status(400).json({ msg: 'Please enter all fields.' })
+    }
+
     try {
         const newTask = new Task({
-            title: req.body.title,
-            category: req.body.category,
-            description: req.body.description,
-            important: req.body.important
+            title,
+            category,
+            description,
+            important
         })
         res.json(await newTask.save())
     } catch(err) {
-        res.status(400).json({ msg: err.toString() })
+        res.status(400).json({ msg: err.message })
     }
 })
 
 // @route   UPDATE api/tasks
 // @desc    Update a task
-// @access  Public
-Router.post('/:id', async (req, res) => {
+// @access  Private
+Router.post('/:id', auth, async (req, res) => {
+    const {
+        title,
+        category,
+        description,
+        important
+    } = req.body
+
+    if(!title || !category || !description ||
+        typeof important === 'undefined') {
+        return res.status(400).json({ msg: 'Please enter all fields.' })
+    }
+
     try {
         const updateTask = {
-            title: req.body.title,
-            category: req.body.category,
-            description: req.body.description,
-            important: req.body.important
+            title,
+            category,
+            description,
+            important
         }
         res.json(await Task.findOneAndUpdate(
             { _id: req.params.id },
@@ -53,19 +79,25 @@ Router.post('/:id', async (req, res) => {
             { new: true }
         ))
     } catch(err) {
-        res.status(400).json({ msg: err.toString() })
+        res.status(400).json({ msg: err.message })
     }
 })
 
 // @route   DELETE api/tasks
 // @desc    Delete a task
-// @access  Public
-Router.delete('/:id', async (req, res) => {
+// @access  Private
+Router.delete('/:id', auth, async (req, res) => {
+    const { id } = req.body
+
+    if(!id) {
+        return res.status(400).json({ msg: 'Please enter all fields.' })
+    }
+
     try {
         res.json(await Task.findOneAndDelete({ _id: req.params.id }))
     } catch(err) {
-        res.status(400).json({ msg: err.toString() })
+        res.status(400).json({ msg: err.message })
     }
 })
 
-module.exports = Router
+export default Router
